fix: rename dynamic route
