+++ examples/fresh/fresh.gen.ts
@@ -4,12 +4,12 @@
 
 import config from "./deno.json" assert { type: "json" };
 import * as $0 from "./routes/index.tsx";
-import * as $1 from "./routes/messages/id.tsx";
+import * as $1 from "./routes/messages/[id].tsx";
 
 const manifest = {
   routes: {
     "./routes/index.tsx": $0,
-    "./routes/messages/id.tsx": $1,
+    "./routes/messages/[id].tsx": $1,
   },
   islands: {},
   baseUrl: import.meta.url,
